# docs/swagger.yaml
basePath: /
definitions:
  avito_segment.Segment:
    properties:
      slug:
        type: string
    required:
    - slug
    type: object
  avito_segment.UserSegmentsRequest:
    properties:
      add_segments:
        items:
          type: string
        type: array
      remove_segments:
        items:
          type: string
        type: array
    type: object
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for AvitoSegment Application
  title: Avito Segment API
  version: "1.0"
paths:
  /api/segments:
    post:
      consumes:
      - application/json
      description: Create a new segment
      parameters:
      - description: Segment data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito_segment.Segment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/avito_segment.Segment'
      summary: Create segment
      tags:
      - segments
  /api/segments/{slug}:
    delete:
      description: Delete a segment by slug
      parameters:
      - description: Segment slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
      summary: Delete segment
      tags:
      - segments
  /api/users/{user_id}/segments:
    get:
      description: Get segments of a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: Get user segments
      tags:
      - users
    post:
      description: Update segments for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Segments data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito_segment.UserSegmentsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
      summary: Update user segments
      tags:
      - users
swagger: "2.0"